third_party/WebKit/LayoutTests/accessibility/aria-scrollbar-role-expected.txt

scrollbar
scrollbar

This tests that the ARIA scrollbar role works correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS scroller.role is 'AXRole: AXScrollBar'
PASS scroller.intValue is 55
PASS scroller.orientation is 'AXOrientation: AXVerticalOrientation'
PASS scroller.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS scroller.role is 'AXRole: AXScrollBar'
PASS scroller.intValue is 55
PASS scroller.orientation is 'AXOrientation: AXVerticalOrientation'
PASS successfullyParsed is true

TEST COMPLETE
